fix(tidy): eliminate readability-redundant-declaration and narrowing conversion warnings
Move cobalt_stable_sort, cobalt_partition, cobalt_unique declarations
inside the SORT_H header guard to fix readability-redundant-declaration.
Add explicit double casts in benchmark.c current_time_ms() to fix
bugprone-narrowing-conversions from __time_t to double.

+++ src/utils/benchmark.c
@@ -13,7 +13,7 @@ static double current_time_ms(void)
 {
     struct timespec ts;
     clock_gettime(CLOCK_MONOTONIC, &ts);
-    return ts.tv_sec * 1000.0 + ts.tv_nsec / 1000000.0;
+    return (double)ts.tv_sec * 1000.0 + (double)ts.tv_nsec / 1000000.0;
 }
 
 double
